requirements.txt
hyperliquid-python-sdk
python-dotenv
eth-account
requests
